+++ .gitignore
@@ -37,6 +37,7 @@ profile.cov
 
 # IdeaIDE
 .idea
+*.iml
 
 # VS Code
 .vscode
